src/ingest/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
dependencies = [
  "google-cloud-secret-manager>=2.26.0",
  "google-cloud-storage>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
